+++ package.json
@@ -11,8 +11,6 @@
     "migrate": "tsx db/migrate.ts"
   },
   "dependencies": {
-    "@emotion/css": "^11.13.0",
-    "@hookform/resolvers": "^3.9.0",
     "@prisma/client": "5.18.0",
     "@radix-ui/react-dropdown-menu": "^2.1.1",
     "@radix-ui/react-icons": "^1.3.0",
@@ -23,27 +21,42 @@
     "@radix-ui/react-slot": "^1.1.0",
     "@radix-ui/react-tabs": "^1.1.0",
     "@tanstack/react-query": "^5.51.23",
+    "@toast-ui/editor": "^3.2.2",
     "@toast-ui/react-editor": "^3.2.3",
     "@types/bcrypt": "^5.0.2",
+    "@types/react-draft-wysiwyg": "^1.13.8",
     "@types/validator": "^13.12.0",
     "bcrypt": "^5.1.1",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "commands": "link:@uiw/react-md-editor/commands",
+    "create-context-state": "^2.0.3",
+    "create-plate-ui": "link:@/lib/create-plate-ui",
+    "custom-markdown-editor": "^0.0.0",
     "is-hotkey": "^0.2.0",
+    "lexical": "^0.17.0",
     "markdown-to-jsx": "^7.4.7",
+    "mdast-util-directive": "^3.0.0",
     "next": "14.2.5",
     "next-auth": "^4.24.7",
     "next-remove-imports": "^1.0.12",
     "next-themes": "^0.3.0",
     "postgres": "^3.4.4",
     "react": "^18.3.1",
+    "react-dnd": "^16.0.1",
+    "react-dnd-html5-backend": "^16.0.1",
     "react-dom": "^18.3.1",
+    "react-draft-wysiwyg": "^1.15.0",
     "react-hook-form": "^7.52.2",
     "react-markdown": "^9.0.1",
+    "react-simple-code-editor": "^0.14.1",
+    "react-simple-wysiwyg": "^3.1.1",
     "redis": "^4.7.0",
+    "refractor": "^4.8.1",
+    "remirror": "^3.0.1",
     "slate": "^0.103.0",
     "slate-history": "^0.100.0",
+    "slate-hyperscript": "^0.100.0",
     "slate-react": "^0.108.0",
     "sonner": "^1.5.0",
     "tailwind-merge": "^2.5.2",
@@ -55,6 +68,7 @@
   },
   "devDependencies": {
     "@tanstack/eslint-plugin-query": "^5.51.15",
+    "@types/is-hotkey": "^0.1.10",
     "@types/node": "^20.14.15",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
