+++ src/components/markdown-editor/something.tsx
@@ -1,0 +1,8 @@
+import React, { Component } from 'react';
+import { Editor } from 'react-draft-wysiwyg';
+import 'react-draft-wysiwyg/dist/react-draft-wysiwyg.css';
+
+
+const EditorComponent = () => <Editor />
+
+export default EditorComponent
